requirements.txt
django<1.11
djangorestframework
markdown
django-filter
django-extensions
django-jsonfield
django-summernote
django-taggit
requests
pytz
# for test
pytest
pytest-django
django-test-plus
model_mommy
pytest-cov
